Move baseline methods to filter and delete baseline module.
