+++ cleanlab/__init__.py
@@ -5,4 +5,3 @@ from . import count
 from . import noise_generation
 from . import filter
 from . import util
-from . import baseline_methods
