+++ cleanlab/filter.py
@@ -25,12 +25,12 @@
 
 from __future__ import (
     print_function, absolute_import, division, unicode_literals, with_statement)
+
+from sklearn.metrics import confusion_matrix
 from sklearn.preprocessing import MultiLabelBinarizer
 import multiprocessing
 from multiprocessing.sharedctypes import RawArray
 import sys
-import os
-import time
 
 from cleanlab.rank import order_label_errors
 from cleanlab.util import (value_counts, round_preserving_row_totals,
@@ -40,6 +40,8 @@ import numpy as np
 # tqdm is a module used to print time-to-complete when multiprocessing is used.
 # This module is not necessary, and therefore is not a package dependency, but 
 # when installed it improves user experience for large datasets.
+from count import calibrate_confident_joint
+
 try:
     import tqdm
 
@@ -555,3 +557,73 @@ def reduce_prune_counts(prune_count_matrix, frac_noise=1.0):
     return new_mat.astype(int)
 
 
+def baseline_argmax(psx, labels, multi_label=False):
+    """This is the simplest baseline approach. Just consider
+    anywhere argmax != s as a label error.
+
+    Parameters
+    ----------
+    labels : np.array
+        A discrete vector of noisy labels, i.e. some labels may be erroneous.
+
+    psx : np.array (shape (N, K))
+        P(label=k|x) is a matrix with K (noisy) probabilities for each of the
+        N examples x. This is the probability distribution over all K classes,
+        for each example, regarding whether the example has label s==k P(s=k|x).
+        psx should have been computed using 3 (or higher) fold cross-validation.
+
+    multi_label : bool
+        Set to True if s is multi-label (list of lists, or np.array of np.array)
+
+    Returns
+    -------
+        A boolean mask that is true if the example belong
+        to that index is label error."""
+
+    if multi_label:
+        return np.array([np.argsort(psx[i]) == j for i, j in enumerate(labels)])
+    return np.argmax(psx, axis=1) != np.asarray(labels)
+
+
+def baseline_argmax_confusion_matrix(
+    psx,
+    labels,
+    calibrate=True,
+    prune_method='prune_by_noise_rate',
+):
+    """This is a baseline approach. That uses the confusion matrix
+    of argmax(psx) and s as the confident joint and then uses cleanlab
+    (confident learning) to find the label errors using this matrix.
+
+    This method does not support multi-label labels.
+
+    Parameters
+    ----------
+
+    labels : np.array
+        A discrete vector of noisy labels, i.e. some labels may be erroneous.
+
+    psx : np.array (shape (N, K))
+        P(label=k|x) is a matrix with K (noisy) probabilities for each of the
+        N examples x. This is the probability distribution over all K classes,
+        for each example, regarding whether the example has label s==k P(s=k|x).
+        psx should have been computed using 3 (or higher) fold cross-validation.
+
+    calibrate : bool
+        Set to True to calibrate the confusion matrix created by pred != given labels.
+        This calibration just makes
+
+    Returns
+    -------
+        A boolean mask that is true if the example belong
+        to that index is label error."""
+
+    confident_joint = confusion_matrix(np.argmax(psx, axis=1), labels).T
+    if calibrate:
+        confident_joint = calibrate_confident_joint(confident_joint, labels)
+    return get_noise_indices(
+        s=labels,
+        psx=psx,
+        confident_joint=confident_joint,
+        prune_method=prune_method,
+    )
